Class room automation python file
The code is supposed to be run in Raspbian on a Raspberry pi. The camera used was a Raspberrry NoIR camera in order to eliminate most of the low light environment problems. The number of fans/lights were predefined and hard-coded. The face detection algorithm uses Haar Cascades provided by OpenCV for face detection. A more robust system can apply multiple cascades to not only detect frontal faces but also faces from different angles, but the computational load would be too much for a Raspberry Pi 2 and the resulting delay between outputs would render the system impractical for real-time applications.